requirements.txt
Django==1.8
django-braces==1.8.1
django-filter==0.10.0
django-gcm==1.2.0
django-oauth-toolkit==0.9.0
django-rest-auth==0.4.0
django-rest-framework-social-oauth2==1.0.2
django-tastypie==0.13.3
djangorestframework==3.1.3
MarkdownWriter==1.1
MySQL-python==1.2.5
oauthlib==1.0.1
PyJWT==1.4.0
python-dateutil==2.5.1
python-mimeparse==1.5.1
python-openid==2.2.5
python-social-auth==0.2.13
pytz==2016.1
requests==2.9.1
requests-oauthlib==0.6.0
six==1.9.0
